requirements.txt
# ==============================================================================
# GenAI-RAG-EEG: Requirements File
# ==============================================================================
#
# Title: EEG-Based Stress Classification using GenAI-RAG Architecture
# Authors: [Your Name], [Co-authors]
# Version: 1.0.0
# Python Version: 3.8+ (Tested on 3.8, 3.9, 3.10, 3.11)
#
# Installation:
#   pip install -r requirements.txt
#
# For GPU support (CUDA):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# ==============================================================================

# =============================================================================
# CORE DEEP LEARNING FRAMEWORK
# =============================================================================
# PyTorch: Primary deep learning framework for model implementation
# - Used for: CNN, LSTM, Attention mechanisms, training loops
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# =============================================================================
# TRANSFORMER MODELS (Text Encoder)
# =============================================================================
# Hugging Face Transformers: Pre-trained language models
# - Used for: Sentence-BERT text encoding, tokenization
# - Model: all-MiniLM-L6-v2 (22.7M parameters, 384-dim embeddings)
transformers>=4.30.0
sentence-transformers>=2.2.0

# =============================================================================
# SCIENTIFIC COMPUTING
# =============================================================================
# NumPy: Numerical computing foundation
# - Used for: Array operations, signal processing, data manipulation
numpy>=1.24.0

# SciPy: Scientific algorithms
# - Used for: Signal filtering, statistical tests, spectral analysis
scipy>=1.10.0

# Pandas: Data manipulation and analysis
# - Used for: Dataset loading, CSV handling, data preprocessing
pandas>=2.0.0

# Scikit-learn: Machine learning utilities
# - Used for: Train/test split, cross-validation, metrics, preprocessing
scikit-learn>=1.2.0

# =============================================================================
# EEG SIGNAL PROCESSING
# =============================================================================
# MNE-Python: EEG/MEG analysis toolkit
# - Used for: EEG preprocessing, filtering, artifact removal
# - Features: ICA, bandpass filtering, epoch extraction
mne>=1.4.0

# =============================================================================
# VISUALIZATION
# =============================================================================
# Matplotlib: Primary plotting library
# - Used for: Training curves, confusion matrices, attention maps
matplotlib>=3.7.0

# Seaborn: Statistical visualization
# - Used for: Heatmaps, distribution plots, correlation matrices
seaborn>=0.12.0

# Plotly: Interactive visualizations
# - Used for: Interactive dashboards, 3D plots
plotly>=5.14.0

# =============================================================================
# VECTOR DATABASE (RAG Retrieval)
# =============================================================================
# FAISS: Facebook AI Similarity Search
# - Used for: Efficient similarity search in embedding space
# - Features: GPU acceleration, approximate nearest neighbors
faiss-cpu>=1.7.4
# For GPU: faiss-gpu>=1.7.4

# ChromaDB: Alternative vector database
# - Used for: Persistent vector storage, metadata filtering
chromadb>=0.4.0

# =============================================================================
# LANGUAGE MODEL INTEGRATION (RAG Generation)
# =============================================================================
# OpenAI: GPT models for explanation generation
# - Used for: Natural language explanation generation
openai>=1.0.0

# LangChain: LLM orchestration framework
# - Used for: RAG pipeline, prompt templates, chain management
langchain>=0.1.0
langchain-community>=0.0.10

# =============================================================================
# WEB FRAMEWORK (Optional: Demo Application)
# =============================================================================
# Flask: Lightweight web framework
# - Used for: REST API, web demo interface
flask>=2.3.0
flask-cors>=4.0.0

# Streamlit: Data app framework (alternative UI)
streamlit>=1.28.0

# =============================================================================
# CONFIGURATION & UTILITIES
# =============================================================================
# PyYAML: Configuration file parsing
# - Used for: Model configs, hyperparameters
pyyaml>=6.0

# python-dotenv: Environment variable management
# - Used for: API keys, database credentials
python-dotenv>=1.0.0

# tqdm: Progress bars
# - Used for: Training loops, data loading
tqdm>=4.65.0

# Rich: Beautiful terminal output
# - Used for: Logging, debugging, status updates
rich>=13.0.0

# =============================================================================
# EXPERIMENT TRACKING
# =============================================================================
# Weights & Biases: Experiment tracking
# - Used for: Hyperparameter logging, model versioning
wandb>=0.15.0

# TensorBoard: Training visualization
# - Used for: Loss curves, model graphs
tensorboard>=2.13.0

# =============================================================================
# IMAGE PROCESSING (for figures)
# =============================================================================
# Pillow: Image processing
# - Used for: Figure generation, DPI handling
Pillow>=9.5.0

# =============================================================================
# TESTING & CODE QUALITY
# =============================================================================
# pytest: Testing framework
pytest>=7.3.0

# black: Code formatting
black>=23.3.0

# flake8: Linting
flake8>=6.0.0

# mypy: Type checking
mypy>=1.3.0

# =============================================================================
# OPTIONAL: DATABASE ORM
# =============================================================================
# SQLAlchemy: Database ORM
# - Used for: Experiment metadata, results storage
sqlalchemy>=2.0.0

# =============================================================================
# OPTIONAL: GPU ACCELERATION (install separately)
# =============================================================================
# NVIDIA CUDA Toolkit required for GPU support
# cupy-cuda11x>=12.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x
